Fix JavaDoc comments and bugs
